requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database - Turso LibSQL
libsql-client==0.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==43.0.1

# HTTP Client for LLM providers
httpx==0.27.2
aiohttp==3.10.10

# LLM Provider SDKs
anthropic==0.39.0
openai==1.54.3
google-generativeai==0.8.3

# Payment Processing
PyJWT==2.9.0
requests==2.32.3

# Utilities
pydantic==2.9.2
pydantic-settings==2.6.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Monitoring & Logging
structlog==24.4.0
